refactor(components/atom/card): add import file extensions
